+++ src/api/swap.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { getSwap } from './swap.js'
+import { type SwapRequest, getSwap } from './swap.js'
 
 const baseSwapRequest = {
   chainId: 1,
@@ -7,9 +7,8 @@ const baseSwapRequest = {
   tokenOut: '0x6B3595068778DD592e39A122f4f5a5cF09C90fE2',
   amount: 1000000000000000000n,
   maxSlippage: 0.005,
-  gasPrice: 8563095715n,
   sender: '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045',
-} as const
+} as const satisfies SwapRequest
 
 describe('getSwap', () => {
   it('should return a swap when recipient is included', async () => {
@@ -19,6 +18,9 @@ describe('getSwap', () => {
     })
 
     expect(result).include({ status: 'Success' })
+    if (result.status === 'Success') {
+      expect(result.tx.gasPrice).toBeGreaterThan(0)
+    }
   })
 
   it('should include a swap with tx.gas when simulate is true', async () => {
@@ -33,7 +35,7 @@ describe('getSwap', () => {
     }
   })
 
-  it.skip('should include a swap when url is set to staging true', async () => {
+  it.skip('should return a swap when url is set to staging', async () => {
     const result = await getSwap({
       ...baseSwapRequest,
       baseUrl: 'https://staging.sushi.com',